Rename package Drop token from properties Change photo size to min available
