--- telegram-bot/src/main/java/com/emc/teamstv/telegrambot/model/BotCommandsEnum.java
@@ -1,23 +1,0 @@
-package com.emc.teamstv.telegrambot.model;
-
-/**
- * Class contains list of available bot commands
- *
- * @author talipa
- */
-
-public enum BotCommandsEnum {
-
-  LIST("/list");
-
-  private final String name;
-
-  BotCommandsEnum(String name) {
-    this.name = name;
-  }
-
-
-  public String getName() {
-    return name;
-  }
-}
