--- telegram-bot/src/main/java/com/emc/teamstv/telegrambot/services/IdGenerator.java
@@ -1,13 +1,0 @@
-package com.emc.teamstv.telegrambot.services;
-
-/**
- * Service provides uniq ID values
- *
- * @param <T> type of id value
- * @author talipa
- */
-
-public interface IdGenerator<T> {
-
-  T getUniq();
-}
