+++ telegram-bot/src/main/java/com/teamstv/telegrambot/handlers/CallbackHandler.java
@@ -1,0 +1,32 @@
+package com.teamstv.telegrambot.handlers;
+
+import com.teamstv.telegrambot.model.ButtonNameEnum;
+import com.teamstv.telegrambot.model.Photo;
+import com.teamstv.telegrambot.services.TransferService;
+import java.util.Optional;
+import org.telegram.telegrambots.meta.api.interfaces.BotApiObject;
+import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
+
+public abstract class CallbackHandler extends Handler {
+
+  protected CallbackHandler(
+      TransferService<Integer, Photo> transferService) {
+    super(transferService);
+  }
+
+  @Override
+  final Optional<? extends BotApiObject> getContent() {
+    return Optional.ofNullable(update.getCallbackQuery());
+  }
+
+  final int getTransferID(ButtonNameEnum nameEnum) {
+    return getContent().map(
+        c -> ((CallbackQuery) c).getData().replace(nameEnum.getData(), "")
+    ).map(Integer::valueOf).orElse(Integer.MAX_VALUE);
+  }
+
+  final Optional<Photo> getPhotoModel(ButtonNameEnum nameEnum) {
+    return transferService.get(getTransferID(nameEnum));
+  }
+
+}
