+++ telegram-bot/src/main/java/com/teamstv/telegrambot/services/IdGenerator.java
@@ -1,0 +1,13 @@
+package com.teamstv.telegrambot.services;
+
+/**
+ * Service provides uniq ID values
+ *
+ * @param <T> type of id value
+ * @author talipa
+ */
+
+public interface IdGenerator<T> {
+
+  T getUniq();
+}
